pyproject.toml
[project]
name = "fractal-printer"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "numpy (>=2.3.2,<3.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "ipykernel (>=6.30.1,<7.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "numpy-quaternion (>=2024.0.10,<2025.0.0)",
    "numpy-stl (>=3.2.0,<4.0.0)",
    "scikit-image (>=0.25.2,<0.26.0)",
    "tqdm (>=4.67.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
